Migrate filter tests WIP
